Remove min/max macros and make template based in coreclr

+++ src/coreclr/vm/profdetach.cpp
@@ -447,8 +447,8 @@ void ProfilingAPIDetach::SleepWhileProfilerEvacuates(ProfilerDetachInfo *pDetach
 
     // ...but keep it in bounds!
     ui64SleepMilliseconds = min(
-        max(ui64SleepMilliseconds, s_dwMinSleepMs),
-        s_dwMaxSleepMs);
+        max(ui64SleepMilliseconds, (ULONGLONG)s_dwMinSleepMs),
+        (ULONGLONG)s_dwMaxSleepMs);
 
     // At this point it's safe to cast ui64SleepMilliseconds down to a DWORD since we
     // know it's between s_dwMinSleepMs & s_dwMaxSleepMs
